Add AutoLoginActivity for seamless user login experience
